構造体の定義を types.go へ移動
